requirements.txt
flask==2.1.2
Jinja2==3.1.2 
MarkupSafe==2.1.1 
Werkzeug==2.1.2 
click==8.1.3
importlib-metadata==4.11.3
itsdangerous==2.1.2 
zipp==3.8.0
python-dotenv==0.20.0
selenium==4.1.5
pycountry==22.3.5